# server/api-doc.json
{
  "openapi": "3.0.0",
  "info": {
    "version": "1.0",
    "title": "HARDWARESTORE-APP(REST API)",
    "contact": {}
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "localhost"
    }
  ],
...
  "schemes": [
    "https",
    "http"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
...
    "/users/register": {
      "post": {
        "tags": [
          "users"
        ],
        "summary": "Register users",
        "operationId": "Registerusers",
        "consumes": [
          "application/x-www-form-urlencoded"
        ],
        "produces": [
          "application/json; charset=utf-8"
        ],
        "parameters": [
          {
            "name": "name",
            "in": "formData",
            "description": "",
            "required": true,
            "type": "string"
          },
          {
            "name": "email",
            "in": "formData",
            "description": "",
            "required": true,
            "type": "string"
          },
          {
            "name": "password",
            "in": "formData",
            "description": "",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "name": "birthdate",
            "in": "formData",
            "description": "",
            "required": true,
            "type": "string"
          },
          {
            "name": "gender",
            "in": "formData",
            "description": "",
            "required": true,
            "type": "string"
          },
          {
            "name": "avatar",
            "in": "formData",
            "description": "",
            "required": true,
            "type": "string"
          },
          {
            "name": "type",
            "in": "formData",
            "description": "",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "403": {
            "description": "Forbidden",
            "examples": {
              "application/json; charset=utf-8": {
                "message": "Email already Used"
              }
            },
            "headers": {
              "X-Powered-By": {
                "type": "string",
                "default": "Express"
              },
              "Content-Length": {
                "type": "string",
                "default": "32"
              },
              "ETag": {
                "type": "string",
                "default": "W/\"20-Z6GZM7FZ1rCmktPUq117kpNukI8\""
              },
              "Date": {
                "type": "string",
                "default": "Sun, 15 Aug 2021 18:22:57 GMT"
              },
              "Connection": {
                "type": "string",
                "default": "keep-alive"
              },
              "Keep-Alive": {
                "type": "string",
                "default": "timeout=5"
              }
            },
            "schema": {
              "$ref": "#/definitions/emailalreadyused1"
            }
          },
          "500": {
            "description": "Internal Server Error",
            "examples": {
              "application/json; charset=utf-8": {
                "name": "SequelizeValidationError",
                "errors": [
                  {
                    "message": "Validation notEmpty on email failed",
                    "type": "Validation error",
                    "path": "email",
                    "value": "",
                    "origin": "FUNCTION",
                    "instance": {
                      "name": "Bagas",
                      "email": "",
                      "password": "0000",
                      "birthdate": "2021-08-10T11:24:36Z",
                      "gender": "Male",
                      "avatar": "tmp\\my-uploads\\avatar-1629051749097-521740772",
                      "type": "",
                      "updatedAt": "2021-08-15T18:22:29.344Z",
                      "createdAt": "2021-08-15T18:22:29.344Z"
                    },
                    "validatorKey": "notEmpty",
                    "validatorName": "notEmpty",
                    "validatorArgs": [
                      {
                        "message": "TIDAK BOLEH KOSONG"
                      }
                    ],
                    "original": {
                      "validatorName": "notEmpty",
                      "validatorArgs": [
                        {
                          "message": "TIDAK BOLEH KOSONG"
                        }
                      ]
                    }
                  }
                ]
              }
            },
            "headers": {
              "X-Powered-By": {
                "type": "string",
                "default": "Express"
              },
              "Content-Length": {
                "type": "string",
                "default": "632"
              },
              "ETag": {
                "type": "string",
                "default": "W/\"278-iRoqckdqpmN6rtwD+oFh/2NFhIc\""
              },
              "Date": {
                "type": "string",
                "default": "Sun, 15 Aug 2021 18:22:29 GMT"
              },
              "Connection": {
                "type": "string",
                "default": "keep-alive"
              },
              "Keep-Alive": {
                "type": "string",
                "default": "timeout=5"
              }
            },
            "schema": {
              "$ref": "#/definitions/emptyform1"
            }
          }
        },
        "deprecated": false
      }
    },
    "/users/login": {
      "post": {
        "tags": [
          "users"
        ],
        "summary": "Login users",
        "operationId": "Loginusers",
        "consumes": [
          "application/x-www-form-urlencoded"
        ],
        "produces": [
          "application/json; charset=utf-8"
        ],
        "parameters": [
          {
            "name": "email",
            "in": "formData",
            "description": "",
            "required": true,
            "type": "string"
          },
          {
            "name": "password",
            "in": "formData",
            "description": "",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "examples": {
              "application/json; charset=utf-8": {
                "access_token": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpZCI6NywibmFtZSI6IkJhZ2FzIiwiZW1haWwiOiJiYWdhc0BnbWFpbC5jb20iLCJzYWx0IjoiJDJiJDEwJHcxeTMvdVR1NmxDSXcvdmR6dEJ3by5YNTJXc0pjNHdtYXhGNFhpTU9tM0czUVhOb0lQUXBTIiwiYmlydGhkYXRlIjoiMjAyMS0wOC0xMFQxMToyNDozNi4wMDBaIiwiZ2VuZGVyIjoiTWFsZSIsImF2YXRhciI6InRtcFxcbXktdXBsb2Fkc1xcYXZhdGFyLTE2MjkwNTE3NzI5ODYtMzIxMTA4NDg2IiwidHlwZSI6IiIsImlhdCI6MTYyOTA1MjQ5Mn0.y1fFe-WkO08e_oXYiyCGgZ0n9IRxGQuAYY2TlVucg1I"
              }
            },
            "headers": {
              "X-Powered-By": {
                "type": "string",
                "default": "Express"
              },
              "Content-Length": {
                "type": "string",
                "default": "447"
              },
              "ETag": {
                "type": "string",
                "default": "W/\"1bf-cFA46GSKf/+9VLnv+PFGJUj9LiI\""
              },
              "Date": {
                "type": "string",
                "default": "Sun, 15 Aug 2021 18:34:52 GMT"
              },
              "Connection": {
                "type": "string",
                "default": "keep-alive"
              },
              "Keep-Alive": {
                "type": "string",
                "default": "timeout=5"
              }
            },
            "schema": {
              "$ref": "#/definitions/loginuserssuccess"
            }
          },
          "403": {
            "description": "Forbidden",
            "examples": {
              "application/json; charset=utf-8": {
                "message": "Invalid Password"
              }
            },
            "headers": {
              "X-Powered-By": {
                "type": "string",
                "default": "Express"
              },
              "Content-Length": {
                "type": "string",
                "default": "30"
              },
              "ETag": {
                "type": "string",
                "default": "W/\"1e-dzv476gwnZrSI5yWtCakgolWpd4\""
              },
              "Date": {
                "type": "string",
                "default": "Sun, 15 Aug 2021 18:35:09 GMT"
              },
              "Connection": {
                "type": "string",
                "default": "keep-alive"
              },
              "Keep-Alive": {
                "type": "string",
                "default": "timeout=5"
              }
            },
            "schema": {
              "$ref": "#/definitions/loginusersinvalidpassword1"
            }
          },
          "404": {
            "description": "Not Found",
            "examples": {
              "application/json; charset=utf-8": {
                "message": " Not Found ! "
              }
            },
            "headers": {
              "X-Powered-By": {
                "type": "string",
                "default": "Express"
              },
              "Content-Length": {
                "type": "string",
                "default": "27"
              },
              "ETag": {
                "type": "string",
                "default": "W/\"1b-ffZOE6IhsAbGRJUNgOHY2Rz/k/c\""
              },
              "Date": {
                "type": "string",
                "default": "Sun, 15 Aug 2021 18:35:22 GMT"
              },
              "Connection": {
                "type": "string",
                "default": "keep-alive"
              },
              "Keep-Alive": {
                "type": "string",
                "default": "timeout=5"
              }
            },
            "schema": {
              "$ref": "#/definitions/registerusersnotfound1"
            }
          }
        },
        "deprecated": false
      }
    },
...
    "/users/update/:id": {
      "put": {
        "tags": [
          "users"
        ],
        "summary": "Update users",
        "operationId": "updateusers",
        "consumes": [
          "application/x-www-form-urlencoded"
        ],
        "produces": [
          "application/json; charset=utf-8"
        ],
        "parameters": [
          {
            "name": "name",
            "in": "formData",
            "description": "",
            "required": true,
            "type": "string"
          },
          {
            "name": "avatar",
            "in": "formData",
            "description": "",
            "required": true,
            "type": "string"
          },
          {
            "name": "email",
            "in": "formData",
            "description": "",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "examples": {
              "application/json; charset=utf-8": [
                1
              ]
            },
            "headers": {
              "X-Powered-By": {
                "type": "string",
                "default": "Express"
              },
              "Content-Length": {
                "type": "string",
                "default": "3"
              },
              "ETag": {
                "type": "string",
                "default": "W/\"3-9imuRLez3P7URNNj5ibt9BHsaag\""
              },
              "Date": {
                "type": "string",
                "default": "Sun, 15 Aug 2021 18:51:37 GMT"
              },
              "Connection": {
                "type": "string",
                "default": "keep-alive"
              },
              "Keep-Alive": {
                "type": "string",
                "default": "timeout=5"
              }
            },
            "schema": {
              "type": "array",
              "items": {
                "type": "integer",
                "format": "int32"
              }
            }
          }
        },
        "deprecated": false
      }
    },
...
  },
  "definitions": {
...
    "Error": {
      "type": "object",
      "required": [
        "instance",
        "message",
        "origin",
        "original",
        "path",
        "type",
        "validatorArgs",
        "validatorKey",
        "validatorName",
        "value"
      ],
      "properties": {
        "message": {
          "type": "string"
        },
        "type": {
          "type": "string"
        },
        "path": {
          "type": "string"
        },
        "value": {
          "type": "string"
        },
        "origin": {
          "type": "string"
        },
        "instance": {
          "$ref": "#/definitions/Instance"
        },
        "validatorKey": {
          "type": "string"
        },
        "validatorName": {
          "type": "string"
        },
        "validatorArgs": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/ValidatorArg"
          }
        },
        "original": {
          "$ref": "#/definitions/Original"
        }
      },
      "title": "Error",
      "example": {
        "message": "Validation notEmpty on email failed",
        "type": "Validation error",
        "path": "email",
        "value": "",
        "origin": "FUNCTION",
        "instance": {
          "name": "Bagas",
          "email": "",
          "password": "0000",
          "birthdate": "2021-08-10T11:24:36Z",
          "gender": "Male",
          "avatar": "tmp\\my-uploads\\avatar-1629051749097-521740772",
          "type": "",
          "updatedAt": "2021-08-15T18:22:29.344Z",
          "createdAt": "2021-08-15T18:22:29.344Z"
        },
        "validatorKey": "notEmpty",
        "validatorName": "notEmpty",
        "validatorArgs": [
          {
            "message": "TIDAK BOLEH KOSONG"
          }
        ],
        "original": {
          "validatorName": "notEmpty",
          "validatorArgs": [
            {
              "message": "TIDAK BOLEH KOSONG"
            }
          ]
        }
      }
    },
    "Instance": {
      "type": "object",
      "required": [
        "avatar",
        "birthdate",
        "createdAt",
        "email",
        "gender",
        "id",
        "name",
        "password",
        "type",
        "updatedAt"
      ],
      "properties": {
        "id": {
          "type": "string"
        },
        "name": {
          "type": "string"
        },
        "email": {
          "type": "string"
        },
        "password": {
          "type": "string"
        },
        "birthdate": {
          "type": "string"
        },
        "gender": {
          "type": "string"
        },
        "avatar": {
          "type": "string"
        },
        "type": {
          "type": "string"
        },
        "updatedAt": {
          "type": "string"
        },
        "createdAt": {
          "type": "string"
        }
      },
      "title": "Instance",
      "example": {
        "name": "Bagas",
        "email": "",
        "password": "0000",
        "birthdate": "2021-08-10T11:24:36Z",
        "gender": "Male",
        "avatar": "tmp\\my-uploads\\avatar-1629051749097-521740772",
        "type": "",
        "updatedAt": "2021-08-15T18:22:29.344Z",
        "createdAt": "2021-08-15T18:22:29.344Z"
      }
    },
    "ValidatorArg": {
      "type": "object",
      "required": [
        "message"
      ],
      "properties": {
        "message": {
          "type": "string"
        }
      },
      "title": "ValidatorArg",
      "example": {
        "message": "TIDAK BOLEH KOSONG"
      }
    },
    "Original": {
      "type": "object",
      "required": [
        "validatorArgs",
        "validatorName"
      ],
      "properties": {
        "validatorName": {
          "type": "string"
        },
        "validatorArgs": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/ValidatorArg"
          }
        }
      },
      "title": "Original",
      "example": {
        "validatorName": "notEmpty",
        "validatorArgs": [
          {
            "message": "TIDAK BOLEH KOSONG"
          }
        ]
      }
    },
    "emptyform1": {
      "type": "object",
      "required": [
        "errors",
        "name"
      ],
      "properties": {
        "name": {
          "type": "string"
        },
        "errors": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Error"
          }
        }
      },
      "title": "emptyform1",
      "example": {
        "name": "SequelizeValidationError",
        "errors": [
          {
            "message": "Validation notEmpty on email failed",
            "type": "Validation error",
            "path": "email",
            "value": "",
            "origin": "FUNCTION",
            "instance": {
              "name": "Bagas",
              "email": "",
              "password": "0000",
              "birthdate": "2021-08-10T11:24:36Z",
              "gender": "Male",
              "avatar": "tmp\\my-uploads\\avatar-1629051749097-521740772",
              "type": "",
              "updatedAt": "2021-08-15T18:22:29.344Z",
              "createdAt": "2021-08-15T18:22:29.344Z"
            },
            "validatorKey": "notEmpty",
            "validatorName": "notEmpty",
            "validatorArgs": [
              {
                "message": "TIDAK BOLEH KOSONG"
              }
            ],
            "original": {
              "validatorName": "notEmpty",
              "validatorArgs": [
                {
                  "message": "TIDAK BOLEH KOSONG"
                }
              ]
            }
          }
        ]
      }
    },
    "emailalreadyused1": {
      "type": "object",
      "required": [
        "message"
      ],
      "properties": {
        "message": {
          "type": "string"
        }
      },
      "title": "emailalreadyused1",
      "example": {
        "message": "Email already Used"
      }
    },
    "loginuserssuccess": {
      "type": "object",
      "required": [
        "access_token"
      ],
      "properties": {
        "access_token": {
          "type": "string"
        }
      },
      "title": "loginuserssuccess",
      "example": {
        "access_token": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpZCI6NywibmFtZSI6IkJhZ2FzIiwiZW1haWwiOiJiYWdhc0BnbWFpbC5jb20iLCJzYWx0IjoiJDJiJDEwJHcxeTMvdVR1NmxDSXcvdmR6dEJ3by5YNTJXc0pjNHdtYXhGNFhpTU9tM0czUVhOb0lQUXBTIiwiYmlydGhkYXRlIjoiMjAyMS0wOC0xMFQxMToyNDozNi4wMDBaIiwiZ2VuZGVyIjoiTWFsZSIsImF2YXRhciI6InRtcFxcbXktdXBsb2Fkc1xcYXZhdGFyLTE2MjkwNTE3NzI5ODYtMzIxMTA4NDg2IiwidHlwZSI6IiIsImlhdCI6MTYyOTA1MjQ5Mn0.y1fFe-WkO08e_oXYiyCGgZ0n9IRxGQuAYY2TlVucg1I"
      }
    },
    "loginusersinvalidpassword1": {
      "type": "object",
      "required": [
        "message"
      ],
      "properties": {
        "message": {
          "type": "string"
        }
      },
      "title": "loginusersinvalidpassword1",
      "example": {
        "message": "Invalid Password"
      }
    },
    "registerusersnotfound1": {
      "type": "object",
      "required": [
        "message"
      ],
      "properties": {
        "message": {
          "type": "string"
        }
      },
      "title": "registerusersnotfound1",
      "example": {
        "message": " Not Found ! "
      }
    },
...
  }
}